seperated tooltips component
